add some comments to method for easy understand.
