+++ metadata/etcd/etcd.go
@@ -77,6 +77,7 @@ func newEtcdMetadata(cfg torus.Config) (torus.MetadataService, error) {
 		uuid = metadata.MakeUUID()
 	} else {
 		uuid, err = metadata.GetUUID(cfg.DataDir)
+		//create an uuid file under the <data-dir>/metadata folder and write one number which generated by uuid into it
 	}
 	if err != nil {
 		return nil, err
@@ -119,6 +120,9 @@ func (e *Etcd) Close() error {
 	return e.Client.Close()
 }
 
+//read the Global metadata string from etcd server with key
+//"/github.com/coreos/torus/meta/globalmetadata"
+//set global value via Unmarshal as a GlobalMetadata struct
 func (e *Etcd) getGlobalMetadata() error {
 	txn := e.Client.Txn(context.Background())
 	resp, err := txn.If(
@@ -142,6 +146,7 @@ func (e *Etcd) getGlobalMetadata() error {
 	return nil
 }
 
+//the only thing is to clone the etcdCtx
 func (e *Etcd) WithContext(ctx context.Context) torus.MetadataService {
 	return &etcdCtx{
 		etcd: e,
@@ -149,12 +154,14 @@ func (e *Etcd) WithContext(ctx context.Context) torus.MetadataService {
 	}
 }
 
+//add one ring listener that will get the new ring when the ring updated
 func (e *Etcd) SubscribeNewRings(ch chan torus.Ring) {
 	e.mut.Lock()
 	defer e.mut.Unlock()
 	e.ringListeners = append(e.ringListeners, ch)
 }
 
+//remove the exist listener
 func (e *Etcd) UnsubscribeNewRings(ch chan torus.Ring) {
 	e.mut.Lock()
 	defer e.mut.Unlock()
@@ -167,6 +174,7 @@ func (e *Etcd) UnsubscribeNewRings(ch chan torus.Ring) {
 
 // Context-sensitive calls
 
+//return the ctx value, the context.Background will be returned when value is nil
 func (c *etcdCtx) getContext() context.Context {
 	if c.ctx == nil {
 		return context.Background()
@@ -174,6 +182,7 @@ func (c *etcdCtx) getContext() context.Context {
 	return c.ctx
 }
 
+//the same as Etcd WithContext to clone one EtcdCtx
 func (c *etcdCtx) WithContext(ctx context.Context) torus.MetadataService {
 	return c.etcd.WithContext(ctx)
 }
@@ -190,6 +199,7 @@ func (c *etcdCtx) UUID() string {
 	return c.etcd.uuid
 }
 
+//add the PeerInfo with lease into etcd server
 func (c *etcdCtx) RegisterPeer(lease int64, p *models.PeerInfo) error {
 	if lease == 0 {
 		return errors.New("no lease")
@@ -231,62 +241,6 @@ func (c *etcdCtx) GetPeers() (torus.PeerInfoList, error) {
 	return torus.PeerInfoList(out), nil
 }
 
-// AtomicModifyFunc is a class of commutative functions that, given the current
-// state of a key's value `in`, returns the new state of the key `out`, and
-// `data` to be returned to the calling function on success, or an `err`.
-//
-// This function may be run multiple times, if the value has changed in the time
-// between getting the data and setting the new value.
-type AtomicModifyFunc func(in []byte) (out []byte, data interface{}, err error)
-
-// TODO(barakmich): Perhaps make this an etcd client library function.
-func (c *etcdCtx) AtomicModifyKey(k []byte, f AtomicModifyFunc) (interface{}, error) {
-	key := string(k)
-	resp, err := c.etcd.Client.Get(c.getContext(), key)
-	if err != nil {
-		return nil, err
-	}
-	var version int64
-	var value []byte
-	if len(resp.Kvs) != 1 {
-		version = 0
-		value = []byte{}
-	} else {
-		kv := resp.Kvs[0]
-		version = kv.Version
-		value = kv.Value
-	}
-	for {
-		newBytes, fval, err := f(value)
-		if err != nil {
-			return nil, err
-		}
-		txn := c.etcd.Client.Txn(c.getContext()).If(
-			etcdv3.Compare(etcdv3.Version(key), "=", version),
-		).Then(
-			etcdv3.OpPut(key, string(newBytes)),
-		).Else(
-			etcdv3.OpGet(key),
-		)
-		resp, err := txn.Commit()
-		if err != nil {
-			return nil, err
-		}
-		if resp.Succeeded {
-			return fval, nil
-		}
-		promAtomicRetries.WithLabelValues(string(key)).Inc()
-		kv := resp.Responses[0].GetResponseRange().Kvs[0]
-		version = kv.Version
-		value = kv.Value
-	}
-}
-
-func BytesAddOne(in []byte) ([]byte, interface{}, error) {
-	newval := BytesToUint64(in) + 1
-	return Uint64ToBytes(newval), newval, nil
-}
-
 func (c *etcdCtx) GetVolumes() ([]*models.Volume, torus.VolumeID, error) {
 	promOps.WithLabelValues("get-volumes").Inc()
 	txn := c.etcd.Client.Txn(c.getContext()).Then(
@@ -311,6 +265,9 @@ func (c *etcdCtx) GetVolumes() ([]*models.Volume, torus.VolumeID, error) {
 	return out, torus.VolumeID(highwater), nil
 }
 
+//get volume from the metadata store with the volume name.
+//get the volume id via request "/github.com/coreos/torus/volumes/<volumename>"
+//get volume information via request "/github.com/coreos/torus/volumes/<volumeid>"
 func (c *etcdCtx) GetVolume(volume string) (*models.Volume, error) {
 	if v, ok := c.etcd.volumesCache[volume]; ok {
 		return v, nil
@@ -343,6 +300,8 @@ func (c *etcdCtx) GetVolume(volume string) (*models.Volume, error) {
 	return v, nil
 }
 
+//query the volume lock status.
+//get value via request "/github.com/coreos/torus/volumemeta/<vol_id>/blocklock"
 func (c *etcdCtx) GetLockStatus(vid uint64) string {
 	resp, err := c.etcd.Client.Get(c.getContext(), MkKey("volumemeta", Uint64ToHex(uint64(vid)), "blocklock"))
 	if err != nil {
@@ -355,6 +314,7 @@ func (c *etcdCtx) GetLockStatus(vid uint64) string {
 	return "in-use"
 }
 
+//query the client lease
 func (c *etcdCtx) GetLease() (int64, error) {
 	resp, err := c.etcd.Client.Grant(c.getContext(), leaseTTL)
 	if err != nil {
@@ -364,22 +324,28 @@ func (c *etcdCtx) GetLease() (int64, error) {
 	return int64(resp.ID), nil
 }
 
+//set the new lease value for client
 func (c *etcdCtx) RenewLease(lease int64) error {
 	lid := etcdv3.LeaseID(lease)
-	resp, err := c.etcd.Client.KeepAliveOnce(c.getContext(), lid)
+	//	resp, err := c.etcd.Client.KeepAliveOnce(c.getContext(), lid)
+	_, err := c.etcd.Client.KeepAliveOnce(c.getContext(), lid)
 	if err != nil {
 		if err == rpctypes.ErrLeaseNotFound {
 			return torus.ErrLeaseNotFound
 		}
 		return err
 	}
-	clog.Tracef("updated lease for %d, TTL %d", resp.ID, resp.TTL)
+	//	clog.Tracef("updated lease for %d, TTL %d", resp.ID, resp.TTL)
 	return nil
 }
+
+//query the ring via EtcdCtx getRing()
 func (c *etcdCtx) GetRing() (torus.Ring, error) {
 	r, _, err := c.getRing()
 	return r, err
 }
+
+//query the ring via request "/github.com/coreos/torus/meta/the-one-ring"
 func (c *etcdCtx) getRing() (torus.Ring, int64, error) {
 	promOps.WithLabelValues("get-ring").Inc()
 	resp, err := c.etcd.Client.Get(c.getContext(), MkKey("meta", "the-one-ring"))
@@ -396,14 +362,17 @@ func (c *etcdCtx) getRing() (torus.Ring, int64, error) {
 	return ring, resp.Kvs[0].Version, nil
 }
 
+//call the Etcd SubscribeNewRings()
 func (c *etcdCtx) SubscribeNewRings(ch chan torus.Ring) {
 	c.etcd.SubscribeNewRings(ch)
 }
 
+//call the Etcd UnsubscribeNewRings()
 func (c *etcdCtx) UnsubscribeNewRings(ch chan torus.Ring) {
 	c.etcd.UnsubscribeNewRings(ch)
 }
 
+//set the new ring value via comparing the version value
 func (c *etcdCtx) SetRing(ring torus.Ring) error {
 	oldr, etcdver, err := c.getRing()
 	if err != nil {
@@ -432,6 +401,9 @@ func (c *etcdCtx) SetRing(ring torus.Ring) error {
 	return torus.ErrAgain
 }
 
+//generate an inode id.
+//transfer the "/github.com/coreos/torus/volumemeta/<vol_id>/inode" into byte array
+//then run byte add 1 then right this value into metadata server
 func (c *etcdCtx) CommitINodeIndex(vid torus.VolumeID) (torus.INodeID, error) {
 	promOps.WithLabelValues("commit-inode-index").Inc()
 	c.etcd.mut.Lock()
@@ -444,6 +416,9 @@ func (c *etcdCtx) CommitINodeIndex(vid torus.VolumeID) (torus.INodeID, error) {
 	return torus.INodeID(newID.(uint64)), nil
 }
 
+//generate an inode id.
+//transfer the "/github.com/coreos/torus/meta/volumeminter" into byte array
+//then write the value generated via adding 1 onto the byte array into metadata
 func (c *etcdCtx) NewVolumeID() (torus.VolumeID, error) {
 	c.etcd.mut.Lock()
 	defer c.etcd.mut.Unlock()
@@ -455,6 +430,7 @@ func (c *etcdCtx) NewVolumeID() (torus.VolumeID, error) {
 	return torus.VolumeID(newID.(uint64)), nil
 }
 
+//get the INode Id from metadata server
 func (c *etcdCtx) GetINodeIndex(vid torus.VolumeID) (torus.INodeID, error) {
 	promOps.WithLabelValues("get-inode-index").Inc()
 	c.etcd.mut.Lock()
